feat: Default value for selecting shelf dropdown
The default value selected in the dropdown is the name of the shelf the book
is currently in. This is so that the user is informed that the book is currently in
a specific shelf. This way, the user will have better experience moving the book
to another shelf. Some minor warnings have also been resolved in this commit
plus an additional component called BookShelfChanger.