refactor: Make BookBubble accept children elements
BookBubble component now can children jsx that will be rendered
inside the bubble. This makes the parent component of BookBubble
responsible for the styling and contents of the book. Because of that,
BookBubble may not be named BookBubble in the next commits
because it is now a more generic and reusable component.