fix: Bubbles with fewer contents render away from the parent hovered
Bubbles with fewer contents render too far away from the hovered parent
because of the hard coded top offset. This happens when the bubble is
supposed to render at the top of its parent. In this commit, the top offset is
calculated based on the height of the bubble for more accurate value.