feat: Star rating for books
The star rating for the books are now working and functional. However, since
the BooksAPI from Udacity does not have an endpoint to receive star ratings
for the books, so the ratings cannot be persisted after hovering off the books.
Meaning the rating will be gone after losing focus on the book that was being
rated.