fix: HoverBubble renders away from direct parent
When hover bubble is moved inside an element that is not absolutely or
relatively positioned, the offsets start from the top left of the screen instead
of the the top left of the parent component. In this commit, BookBubble
gets it's direct parent, and that direct parent must be relative or absolute
in position.